Backend/requirements.txt
fastapi
uvicorn
whisper-openai
librosa
moviepy
transformers
xgboost
scikit-learn
mediapipe
opencv-python
